Desktop/school/microservice/try3/llmMicroService/rag-service/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic>=2.0.0,<3.0.0
langchain>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-google-genai>=1.0.3,<1.1.0
langchain-text-splitters>=0.2.0,<0.3.0
langchain-huggingface>=0.0.1,<0.1.0
langchain-chroma>=0.1.0,<0.2.0
chromadb==0.5.20
pypdf
sentence-transformers
google-generativeai
optimum[onnxruntime]